requirements.txt
Django==5.2.7
environs==14.3.0
pillow==12.0.0
python-dateutil~=2.9.0.post0
djangorestframework==3.15.2
